feat: expand uiController interaction functions I intended to add these features incrementally, but got caught in the flow and the ended up bundled together in this commit. Future refactoring planned to enhance this wip code.
changes made:

- Added colorPlayerShips to visually indicate player ships on the board.
- Implemented handleCellClick to manage player interactions.
This function handles user clicks on the game board, determining the coordinates of the click and delegating the attack logic to the game controller.
It updates the UI based on hit or miss and checks for win conditions.
- Created setupCellClickHandler to attach click event listeners to the game board.
This function ensures user clicks are captured and processed, linking the UI with game logic.
- Added aiMakeAttack for AI's turn in gameplay.
This function allows the AI to make a random attack and updates the UI accordingly