fix: failing checkWin test
This commit refactors the setup in GameController tests to manually place ships instead of relying on the initializeGame method.
Previously, the placeAiShips(); method called in initializeGame interfered with the test's ability to predictably determine if all ships were sunk.
This leading to inconsistent test outcomes. To rectify this I have called the board set ups manually to ensure consistency in testing conditions.